Fixed a bug where grabrc push would fail on a directory if the directory was not already in the repo (would try to mistakenly remove it from the tmp dir).

+++ client/uploader.py
@@ -64,7 +64,8 @@ def save(source_path, options):
                          path_in_repo)
         elif os.path.isdir(source_path):
             # shutil.copytree doesn't overwrite an existing directory
-            shutil.rmtree(path_in_repo)
+            if os.path.isdir(path_in_repo):
+                shutil.rmtree(path_in_repo)
             shutil.copytree(source_path,
                             os.path.join(tmp_repo, basename))
     except IOError, e:
